+++ packages/masonex/debug_paths.dart
@@ -1,0 +1,7 @@
+import 'dart:io';
+import 'test/helpers/get_brick_path.dart';
+
+void main() {
+  print('hello_world: ${getBrickPath('hello_world')}');
+  print('todos: ${getBrickPath('todos')}');
+}