+++ bricks/hello/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.1.0+2
 
-- chore: upgrade to `mason ^0.1.0`
+- chore: upgrade to `masonex ^0.1.0`
 
 # 0.1.0+1